pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Albert5a <albertsoares97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-extensions = "^3.2.3"
django-debug-toolbar = "^5.0.1"
flake8 = "^7.1.1"
whitenoise = "^6.9.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
